Version 2.10 released
Split UNIVERSAL::ExtDirect sub declaration into separate modules for
Perls < 5.12 and 5.12+, which makes RPC::ExtDirect 2.x compatible with
Perl 5.6+ again. With older Perls, ExtDirect attribute handler will be
processed in CHECK phase as it was in 1.x; with 5.12+ the handler is
processed in BEGIN phase to make it compatible with Apache2/mod_perl.
So now we have the best of two worlds and it is no longer necessary to
keep RPC::ExtDirect 1.x for older Perls.